+++ api/src/main/java/pro/mbroker/api/dto/request/CalculatorRequest.java
@@ -26,7 +26,7 @@ public class CalculatorRequest {
 
     private BigDecimal downPayment;
 
-    private Integer maxMonthlyPayment;
+    private BigDecimal maxMonthlyPayment;
 
     private Integer creditTerm;
